backend/requirements.txt
# Carbonly Backend - Python 3.11+
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
greenlet==3.1.1
alembic==1.14.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
argon2-cffi==25.1.0
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.17
httpx==0.28.1
python-dotenv==1.0.1
stripe==11.6.0

# PDF generation
weasyprint==62.3
jinja2==3.1.4
pydyf==0.11.0

# Optional: Google OAuth
authlib==1.3.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Linting & formatting
ruff==0.8.4
black==24.10.0
